+++ src/Kit/Feedback/index.ts
@@ -2,4 +2,4 @@ export {KitAlert} from './Alert';
 export {KitModal} from './Modal';
 export {KitProgress} from './Progress';
 export {KitSnackBarProvider, openKitSnackBar, closeKitSnackBar} from './SnackBar';
-export {useKitNotification} from './Notification';
+export {useKitNotification, KitNotification} from './Notification';