+++ stories/02-Design System/05-DataDisplay/ItemCard/ItemCard.stories.tsx
@@ -8,7 +8,7 @@ import {DisabledItemCardTest} from './test-components/DisabledItemCardTest.tsx';
 
 const meta: Meta<typeof KitItemCard> = {
     component: KitItemCard,
-    //title: 'Design System/DataDisplay/ItemCard',
+    title: 'Deprecated/DataDisplay/ItemCard',
     argTypes: argTypes
 };