+++ stories/02-Design System/06-Feedback/Notification/Notification.stories.tsx
@@ -1,6 +1,12 @@
+import React from 'react';
 import type {Meta, StoryObj} from '@storybook/react';
 import {argTypes} from './data';
 import {FakeNotification, Template} from './Template';
+import {BasicNotificationTest} from './test-components/basicNotificationTest';
+import {SubtitleNotificationTest} from './test-components/subtitleNotificationTest';
+import {ClosableNotificationTest} from './test-components/closableNotificationTest';
+import {ButtonsNotificationTest} from './test-components/buttonsNotificationTest';
+import {FullNotificationTest} from './test-components/fullNotificationTest';
 
 const meta: Meta<typeof FakeNotification> = {
     component: FakeNotification,
@@ -16,4 +22,16 @@ export const Api: Story = {
     parameters: {
         chromatic: {disableSnapshot: true}
     }
+};
+
+export const ChromaticTest: Story = {
+    render: () => (
+        <div style={{display: 'flex', flexDirection: 'column', gap: '10px'}}>
+            <BasicNotificationTest />
+            <ClosableNotificationTest />
+            <SubtitleNotificationTest />
+            <ButtonsNotificationTest />
+            <FullNotificationTest />
+        </div>
+    )
 };